restructure into t_symbol_stack module
